fix(FIX-02): cast monetary sums as bigint to prevent int overflow in financial reports
Change cast(sum(...) as int) to bigint for all halalas monetary aggregates
across the 8 financial reports (trial-balance, periods, pl, balance-sheet,
vat-return, cash-flow, booking-profitability, supplier-profitability).

A 32-bit int caps at ~21.4M SAR; large agencies could overflow. Row-count
casts (count(*), count(distinct)) remain int as they cannot overflow.

balance-sheet and cash-flow already used bigint; no change needed there.

+++ apps/web/src/app/api/accounting/periods/route.ts
@@ -54,8 +54,8 @@ async function createYearEndClosingEntry(
       accountCode:   journalLines.accountCode,
       accountNameAr: journalLines.accountNameAr,
       accountNameEn: journalLines.accountNameEn,
-      totalDebit:    sql<number>`cast(sum(${journalLines.debitHalalas})  as int)`,
-      totalCredit:   sql<number>`cast(sum(${journalLines.creditHalalas}) as int)`,
+      totalDebit:    sql<number>`cast(sum(${journalLines.debitHalalas})  as bigint)`,
+      totalCredit:   sql<number>`cast(sum(${journalLines.creditHalalas}) as bigint)`,
     })
     .from(journalLines)
     .innerJoin(journalEntries, eq(journalLines.entryId, journalEntries.id))
